Inventory Fix and Containers
I have started working on containers. Right now they are w.i.p but they
are in the game. Also inventory has been fixed. I changed the mapping of
some keys. F now picks up things and opens things while G opens and
closes the inventory.

+++ Game/Source/Game/MasterItem.h
@@ -17,7 +17,7 @@ struct FItemDataStruct
 	UTexture2D *Icon;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
-	int32 ItemID;
+	float ItemID;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
 	FString Description;
